Adding the labels to account page
Added the email and username to account page. For some reason
GetUserData is not working right now so the username is hardcoded, but
will fix the method later. Also started work on getting the stats page
not hardcoded, but that was very basic and not much done yet for it.