Getting scores for active quiz working
The business and database layers for all of the methods for getting the
scores of the given active quiz are now all working. Some changes will
need to be made for making the Boxplot. Mainly, we have decieded to
pass in the Statstitcs for it, but it was throwing an error so I had to
comment it out for the time being. I also commented the business and
database layer changes I made.